@@ -113,6 +113,7 @@
     "@babel/preset-env": "^7.28.3",
     "@babel/preset-react": "^7.27.1",
     "@babel/preset-typescript": "^7.27.1",
+    "@eslint/js": "^9.34.0",
     "@fetch-mock/jest": "^0.2.16",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.6.4",
@@ -144,7 +145,7 @@
     "css-loader": "^7.1.2",
     "cssnano": "^7.1.0",
     "devtools-license-check": "^0.9.0",
-    "eslint": "^8.57.1",
+    "eslint": "^9.34.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-import": "^2.32.0",
@@ -157,6 +158,7 @@
     "fake-indexeddb": "^6.1.0",
     "file-loader": "^6.2.0",
     "glob": "^11.0.3",
+    "globals": "^16.3.0",
     "html-webpack-plugin": "^5.6.4",
     "husky": "^4.3.8",
     "jest": "^30.0.5",
