@@ -2,9 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 //
-const { parentPort, workerData } = require('worker_threads');
-const fs = require('fs');
-const vm = require('vm');
+import { parentPort, workerData } from 'worker_threads';
+import fs from 'fs';
+import vm from 'vm';
 
 if (typeof workerData !== 'string') {
   throw new Error(`Please pass a file name using the 'workerData' property.`);
