@@ -9,7 +9,7 @@ class NodeWorker {
   onmessage: ((event: MessageEvent) => unknown) | null;
 
   constructor(file: string) {
-    const worker = new Worker(__dirname + '/node-worker-contents.js', {
+    const worker = new Worker(__dirname + '/node-worker-contents.mjs', {
       workerData: file,
     });
     worker.on('message', this.onMessage);
